UGTs_CNN/requirements.txt
torch==1.5.1
torchvision==0.6.1
pyyaml==5.3.1
pandas==1.2.0
